+++ ArmiesSource/EarthArmy/eHeal.cpp
@@ -2,7 +2,7 @@
 
 
 
-eHeal::eHeal(int id ,int Tj, int Health, int AttackCapacity, int AttackPower, Game* game) :
+eHeal::eHeal(int id ,int Tj, double Health, int AttackCapacity, double AttackPower, Game* game) :
 	unit(id , EH, Tj, Health, AttackCapacity, AttackPower,game)
 {
